+++ send_push.go
@@ -6,15 +6,15 @@ import (
 )
 
 type SendPushRequest struct {
-	MessageData             map[string]interface{} `json:"message_data,omitempty"`
-	TransactionalMessageID  string                 `json:"transactional_message_id,omitempty"`
-	Identifiers             map[string]string      `json:"identifiers"`
-	To                      string                 `json:"to,omitempty"`
-	DisableMessageRetention *bool                  `json:"disable_message_retention,omitempty"`
-	SendToUnsubscribed      *bool                  `json:"send_to_unsubscribed,omitempty"`
-	QueueDraft              *bool                  `json:"queue_draft,omitempty"`
-	SendAt                  *int64                 `json:"send_at,omitempty"`
-	Language                *string                `json:"language,omitempty"`
+	MessageData             map[string]any    `json:"message_data,omitempty"`
+	TransactionalMessageID  string            `json:"transactional_message_id,omitempty"`
+	Identifiers             map[string]string `json:"identifiers"`
+	To                      string            `json:"to,omitempty"`
+	DisableMessageRetention *bool             `json:"disable_message_retention,omitempty"`
+	SendToUnsubscribed      *bool             `json:"send_to_unsubscribed,omitempty"`
+	QueueDraft              *bool             `json:"queue_draft,omitempty"`
+	SendAt                  *int64            `json:"send_at,omitempty"`
+	Language                *string           `json:"language,omitempty"`
 
 	Title         string          `json:"title,omitempty"`
 	Message       string          `json:"message,omitempty"`
@@ -22,7 +22,7 @@ type SendPushRequest struct {
 	Link          string          `json:"link,omitempty"`
 	CustomData    json.RawMessage `json:"custom_data,omitempty"`
 	CustomPayload json.RawMessage `json:"custom_payload,omitempty"`
-	Device        *deviceV2       `json:"custom_device,omitempty"`
+	Device        *Device         `json:"custom_device,omitempty"`
 	Sound         string          `json:"sound,omitempty"`
 }